backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pymupdf
python-docx
scikit-learn
numpy
joblib
